+++ src/machine.h
@@ -64,14 +64,16 @@ class machine final {
     pma_entry &register_pma_entry(pma_entry &&pma);
 
     /// \brief Creates a new PMA entry reflecting a memory range configuration.
+    /// \param description Informative description of PMA entry for use in error messages
     /// \param c Memory range configuration.
     /// \returns New PMA entry (with default flags).
-    static pma_entry make_memory_range_pma_entry(const memory_range_config &c);
+    static pma_entry make_memory_range_pma_entry(const std::string &description, const memory_range_config &c);
 
     /// \brief Creates a new flash drive PMA entry.
+    /// \param description Informative description of PMA entry for use in error messages
     /// \param c Memory range configuration.
     /// \returns New PMA entry with flash drive flags already set.
-    static pma_entry make_flash_drive_pma_entry(const memory_range_config &c);
+    static pma_entry make_flash_drive_pma_entry(const std::string &description, const memory_range_config &c);
 
     /// \brief Creates a new rollup rx buffer PMA entry.
     /// \param c Memory range configuration.