+++ src/shadow-pmas-factory.cpp
@@ -61,7 +61,8 @@ pma_entry make_shadow_pmas_pma_entry(uint64_t start, uint64_t length) {
         false,                      // IW
         PMA_ISTART_DID::shadow_pmas // DID
     };
-    return make_device_pma_entry(start, length, shadow_pmas_peek, &shadow_pmas_driver).set_flags(f);
+    return make_device_pma_entry("shadow PMAs device", start, length, shadow_pmas_peek, &shadow_pmas_driver)
+        .set_flags(f);
 }
 
 } // namespace cartesi