+++ src/uarch-machine.cpp
@@ -20,6 +20,8 @@
 
 namespace cartesi {
 
+using namespace std::string_literals;
+
 const pma_entry::flags uarch_machine::m_rom_flags{
     true,                  // R
     false,                 // W
@@ -49,40 +51,59 @@ uarch_machine::uarch_machine(uarch_config c) : m_s{}, m_c{std::move(c)} {
     }
     // Register memory PMAs
     if (!m_c.ram.image_filename.empty()) {
-        register_pma_entry(make_callocd_memory_pma_entry(PMA_UARCH_RAM_START, m_c.ram.length, m_c.ram.image_filename)
-                               .set_flags(m_ram_flags));
+        register_pma_entry(
+            make_callocd_memory_pma_entry("uarch RAM"s, PMA_UARCH_RAM_START, m_c.ram.length, m_c.ram.image_filename)
+                .set_flags(m_ram_flags));
     } else if (m_c.ram.length > 0) {
-        register_pma_entry(make_callocd_memory_pma_entry(PMA_UARCH_RAM_START, m_c.ram.length).set_flags(m_ram_flags));
+        register_pma_entry(
+            make_callocd_memory_pma_entry("uarch RAM"s, PMA_UARCH_RAM_START, m_c.ram.length).set_flags(m_ram_flags));
     }
 
     if (!m_c.rom.image_filename.empty()) {
-        register_pma_entry(make_callocd_memory_pma_entry(PMA_UARCH_ROM_START, m_c.rom.length, m_c.rom.image_filename)
-                               .set_flags(m_rom_flags));
+        register_pma_entry(
+            make_callocd_memory_pma_entry("uarch ROM"s, PMA_UARCH_ROM_START, m_c.rom.length, m_c.rom.image_filename)
+                .set_flags(m_rom_flags));
     } else if (m_c.rom.length > 0) {
-        register_pma_entry(make_callocd_memory_pma_entry(PMA_UARCH_ROM_START, m_c.rom.length).set_flags(m_rom_flags));
+        register_pma_entry(
+            make_callocd_memory_pma_entry("uarch ROM"s, PMA_UARCH_ROM_START, m_c.rom.length).set_flags(m_rom_flags));
     }
 
-    register_pma_entry(make_empty_pma_entry(0, 0));
+    register_pma_entry(make_empty_pma_entry("uarch sentinel"s, 0, 0));
 }
 
 pma_entry &uarch_machine::register_pma_entry(pma_entry &&pma) {
     if (m_s.pmas.capacity() <= m_s.pmas.size()) { // NOLINT(readability-static-accessed-through-instance)
-        throw std::runtime_error{"too many PMAs"};
+        throw std::runtime_error{"too many PMAs when adding "s + pma.get_description()};
     }
     auto start = pma.get_start();
     if ((start & (PMA_PAGE_SIZE - 1)) != 0) {
-        throw std::invalid_argument{"PMA start must be aligned to page boundary"};
+        throw std::invalid_argument{"start of "s + pma.get_description() + " ("s + std::to_string(start) +
+            ") must be aligned to page boundary of "s + std::to_string(PMA_PAGE_SIZE) + " bytes"s};
     }
     auto length = pma.get_length();
     if ((length & (PMA_PAGE_SIZE - 1)) != 0) {
-        throw std::invalid_argument{"PMA length must be multiple of page size"};
+        throw std::invalid_argument{"length of "s + pma.get_description() + " ("s + std::to_string(length) +
+            ") must be multiple of page size "s + std::to_string(PMA_PAGE_SIZE)};
+    }
+    // Check PMA range, when not the sentinel PMA entry
+    if (!(length == 0 && start == 0)) {
+        if (length == 0) {
+            throw std::invalid_argument{"length of "s + pma.get_description() + " cannot be zero"s};
+        }
+        // Checks if PMA is in addressable range, safe unsigned overflows
+        if (start > PMA_ADDRESSABLE_MASK || (length - 1) > (PMA_ADDRESSABLE_MASK - start)) {
+            throw std::invalid_argument{
+                "range of "s + pma.get_description() + " must use at most 56 bits to be addressable"s};
+        }
     }
     // Range A overlaps with B if A starts before B ends and A ends after B starts
     for (const auto &existing_pma : m_s.pmas) {
         if (start < existing_pma.get_start() + existing_pma.get_length() && start + length > existing_pma.get_start()) {
-            throw std::invalid_argument{"PMA overlaps with existing PMA"};
+            throw std::invalid_argument{"range of "s + pma.get_description() + " overlaps with range of existing "s +
+                existing_pma.get_description()};
         }
     }
+    pma.set_index(static_cast<int>(m_s.pmas.size()));
     m_s.pmas.push_back(std::move(pma));
     return m_s.pmas.back();
 }