+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Optimized Lua API to avoid allocating a uservalue for error messages for every API call.
 - Lower PMA addressable ranges from 64-bit to 56-bit.
 - Changed marchid to 0xd
+- Improved error messages relating to PMAs
 
 ### Fixed